Adding more test cases for back-end APIs and repository classes.
